pyproject.toml
[project]
name = "ai-chat-notion-be"
version = "0.1.0"
description = "OCR backend with FastAPI"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.128.2",
    "uvicorn>=0.40.0",
    "pytesseract",
    "pillow",
    "python-multipart",
    "google-genai>=1.62.0",
    "requests>=2.32.5",
    "notion-client>=2.7.0",
    "python-dotenv>=1.2.1",
]
